Uncommented problematic code, moved CPU and Opcodes into their own respective files.
